Implemented maximum percentage of cores allocation policy
By specifying a number between 0.0 and 1.0 in the property
spark.dynamicAllocation.maxPercExecutors, each application will not use
more than the specified percentage of cores among all the available
cores in the cluster.
Note that adding nodes to the cluster in a dynamic way will not make
the allocated nodes increase (at the moment).